% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_planing.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the weights of an ATRP reaction}
\usage{
calc_weights(db, mass_product = 1, conversation = 1,
  monomere = "MMA", initiator = "EBriB", catalyst = "CuBr",
  ligand = "PMDETA", ratio_m_i = 200, ratio_cat_i = 1.5,
  ratio_lig_i = 3)
}
\arguments{
\item{db}{A tibble. A chemical database with e.g. molecular weights. Usually  ---- insert DB ---.}

\item{mass_product}{A numeric. The desired mass in g.}

\item{conversation}{A numeric.}

\item{monomere}{A string.}

\item{initiator}{A string.}

\item{catalyst}{A string.}

\item{ligand}{A string.}

\item{ratio_m_i}{A numeric.}

\item{ratio_cat_i}{A numeric.}

\item{ratio_lig_i}{A numeric.}
}
\description{
Calculate the weights of an ATRP reaction
}
