% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_planing.R
\name{DBCP_sur_DPs}
\alias{DBCP_sur_DPs}
\title{Calculate the degree of polymerisation of the blocks of a DBCP-surfactant}
\usage{
DBCP_sur_DPs(M_tot, hlb, M_sol_mono, M_insol_mono)
}
\arguments{
\item{M_tot}{A numeric. The total molecular weigth of the surfactant in
g/mol.}

\item{hlb}{A numeric between 0 and 1. The hydrophilic balance, or more exact
the proportion of the molecular weight of the water soluble block on the
total molecular weight of the surfactant.}

\item{M_sol_mono}{A numeric. The molecular weight of the monomer forming the
water soluble block in g/mol.}

\item{M_insol_mono}{A numeric. The molecular weight of the monomer forming
the water insuble block in g/mol.}
}
\description{
Calculate the degree of polymerisation of the blocks of a DBCP-surfactant
}
