% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolymer.R
\docType{package}
\name{ggpolymer}
\alias{ggpolymer}
\alias{ggpolymer-package}
\title{ggpolymer: A package containing a collection of useful functions for everyday work in the Gröschel polymer lab.}
\description{
Until now the package helps with:
planing ATRP reactions and processing gpc data
}
\details{
In future i plan on adding: DLS and NMR data processing, and at some point a ATRP database, as well as a shiny app for easy use.
}
