% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_pretty_plots.R
\name{mwd_pp}
\alias{mwd_pp}
\title{Function to import gpc mwd data as a tidy tibble, mostly used for pplot_gpc_mwd}
\usage{
mwd_pp(dir, curve_names)
}
\arguments{
\item{dir}{A string. The dir.}

\item{curve_names}{A vector of strings. The names of the GPC curves".}
}
\description{
Function to import gpc mwd data as a tidy tibble, mostly used for pplot_gpc_mwd
}
