% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{plotting function which is able to plot different GPC spectra depending on the chosen data_type}
\usage{
plot_traces(traces, data_type = "elu", normalized_RI = FALSE,
  durations = FALSE, ...)
}
\arguments{
\item{traces}{A list of gpc data tibbles (best used with traces_in_dir(),}

\item{data_type}{A string. /"raw/", "/elu/" or /"mwd/".}

\item{durations}{A vector (numeric or string) of reaction durations at sample
time.}
}
\description{
plotting function which is able to plot different GPC spectra depending on the chosen data_type
}
\seealso{
\code{\link{traces_in_dir}})
}
