% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_traces.R
\name{plot_traces_in_dir}
\alias{plot_traces_in_dir}
\title{Function to plot all the gpc plots of interest in a dirctory}
\usage{
plot_traces_in_dir(dir, data_type = "elu", normalized_RI = FALSE, ...)
}
\arguments{
\item{dir}{A string. The dir.}

\item{data_type}{A string. "elu", "raw", or "mwd".}

\item{normalized_RI}{TRUE or FALSE. Use normalized RI data or not.}
}
\description{
Function to plot all the gpc plots of interest in a dirctory
}
