% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_pretty_plots.R
\name{pplot_gpc_mwd}
\alias{pplot_gpc_mwd}
\title{Function to pretty plot all gpc mwd curves in a dirctory}
\usage{
pplot_gpc_mwd(mwd)
}
\arguments{
\item{mwd}{A tibble. Usually created by the mwd_pp function.}
}
\description{
Function to pretty plot all gpc mwd curves in a dirctory
}
