% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot refractive index vs elution volume from a list of gpc data tibbles}
\usage{
plot_traces(traces, durations = FALSE, xlim = c(10, 45), ylim = c(0,
  100), normalized_RI = FALSE)
}
\arguments{
\item{traces}{A list of gpc data tibbles (best used with traces_in_dir(),}

\item{durations}{a vector (numeric or string) of reaction durations at sample
time.}
}
\description{
Plot refractive index vs elution volume from a list of gpc data tibbles
}
\seealso{
\code{\link{traces_in_dir}})
}
