% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_traces.R
\name{traces_in_dir}
\alias{traces_in_dir}
\title{Read all gpc ascii files in a directory}
\usage{
traces_in_dir(dir, data_type = "elu")
}
\arguments{
\item{dir}{A string.}
}
\value{
A list of gpc data tibbles as elements.
}
\description{
Read all gpc ascii files in a directory
}
